GreenQrGen/requirements.txt
qrcode
Pillow
argparse
base45
flynn
cose